+++ frontend/interfaces/Data.tsx
@@ -1,0 +1,7 @@
+import { FilterTypes } from "@/interfaces/Filters"
+import { NewEvent } from "@/interfaces/Events"
+
+export interface Data {
+  eventData: Array<NewEvent>,
+  filters: FilterTypes
+}
